+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All user visible changes to this project will be documented in this file. This p
 ### Added
 
 - `RtpTransceiverInit.sendEncodings` field with `SendEncodingParameters`. ([#125])
+- `MediaStreamTrack.height()` and `MediaStreamTrack.width()` methods. ([#129])
 - `RtpParameters` class, `RtpSender.getParameters()` and `RtpSender.setParameters()` methods. ([#135])
 
 ### Changed
@@ -34,6 +35,7 @@ All user visible changes to this project will be documented in this file. This p
 [#123]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/123
 [#124]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/124
 [#125]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/125
+[#129]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/129
 [#133]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/133
 [#135]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/135
 [#136]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/136